@@ -8,16 +8,15 @@
     "serve": "nodemon ./dist/server.js",
     "watch": "tsc -w",
     "transpile": "tsc",
-    "seed": "npm run transpile && node ./dist/src/db/seed.js",
     "lint": "tslint -c tslint.json -p tsconfig.json"
   },
   "engines": {
-    "node": ">=8.0"
+    "node": ">=16.19.1"
   },
   "keywords": [
     "Node",
     "Typescript",
-    "postgresql",
+    "mongodb",
     "javascript"
   ],
   "repository": {
@@ -32,30 +31,27 @@
   "license": "",
   "devDependencies": {
     "@types/dotenv": "^8.2.0",
-    "@types/express": "^4.17.11",
+    "@types/express": "^5.0.0",
     "@types/helmet": "4.0.0",
     "@types/joi": "^17.2.3",
-    "@types/node": "^15.0.1",
-    "@types/sequelize": "^4.28.9",
+    "@types/node": "^22.10.5",
     "@types/winston": "^2.4.4",
     "concurrently": "^6.0.2",
-    "nodemon": "^2.0.7",
+    "nodemon": "^3.1.9",
     "tslint": "^6.1.2",
-    "typescript": "^4.2.4"
+    "typescript": "^5.7.2"
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
-    "dotenv": "^8.6.0",
+    "dotenv": "^16.4.7",
     "express": "^4.21.2",
-    "express-rate-limit": "^5.2.6",
-    "helmet": "^4.5.0",
-    "husky": "^6.0.0",
+    "express-rate-limit": "^7.5.0",
+    "helmet": "^8.0.0",
+    "husky": "^9.1.7",
     "joi": "^17.13.3",
     "mongoose": "^8.9.3",
     "morgan": "^1.10.0",
-    "pg": "^8.6.0",
-    "sequelize": "^6.6.2",
-    "winston": "^3.3.3"
+    "winston": "^3.17.0"
   },
   "hooks": {
     "pre-commit": "npm run lint"
