@@ -1,4 +1,4 @@
-import * as RateLimit from 'express-rate-limit';
+import { rateLimit } from 'express-rate-limit';
 
 const env = process.env.NODE_ENV || 'dev';
 const rateLimitRequest = Number(process.env.RATE_LIMIT_TIME) || 15;
@@ -6,17 +6,17 @@ const rateLimitTime = Number(process.env.RATE_LIMIT_REQUEST) || 100;
 
 export default () => {
   if (env === 'production') {
-    return new RateLimit({
+    return rateLimit({
       windowMs: rateLimitTime * 60 * 1000, // 15 minutes
       max: rateLimitRequest, // limit each IP to 30 requests per windowMs
-      delayMs: 0,
-      handler: 'Rate limt exceeded, please try again later some time.',
+      handler: (req, res, next, options) =>
+        res.status(options.statusCode).send('Rate limt exceeded, please try again later some time.'),
     });
   }
-  return new RateLimit({
+  return rateLimit({
     windowMs: 5 * 60 * 1000, // 5 minutes
     max: 3000, // limit each IP to 3000 requests per windowMs
-    delayMs: 0,
-    handler: 'Rate limt exceeded, please try again later some time.',
+    handler: (req, res, next, options) =>
+      res.status(options.statusCode).send('Rate limt exceeded, please try again later some time.'),
   });
 };
