@@ -1,4 +1,4 @@
-import * as Joi from "joi";
+import Joi from "joi";
 import { apiErrorHandler } from "../handlers/errorHandler";
 import { pick } from "../utils/pick";
 
